+++ packages/css/src/components/columns-list/columns-list.stories.ts
@@ -1,0 +1,16 @@
+import { storiesOfColumnsList } from '@dso-toolkit/sources';
+import { storiesOf } from '@storybook/web-components';
+
+import { columnsListTemplate } from './columns-list.template';
+import readme from './readme.md';
+
+storiesOfColumnsList(
+  {
+    storiesOf,
+    module,
+    readme
+  },
+  {
+    columnsListTemplate
+  }
+);