+++ packages/sources/src/components/columns-list/columns-list.models.ts
@@ -1,0 +1,7 @@
+export interface ColumnsList {
+  listItems: ColumnsListItem[];
+}
+
+export interface ColumnsListItem {
+  listItem: string;
+}