+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * **BREAKING: core:** Tree View items dienen een "selected" en een "active" state te ondersteunen ([#1349](https://github.com/dso-toolkit/dso-toolkit/issues/1349))
 
+### Changed
+* **css + dso-toolkit:** Lange lijsten in definition list tonen als kolommen ([#1378](https://github.com/dso-toolkit/dso-toolkit/issues/1378))
+
 ## 32.2.0
 
 ### Added